+++ Mechanics/SL/Common.lua
@@ -13,10 +13,14 @@ local SLSpells = {
     [343520] = true, -- Storming
     [226512] = true, -- Sanguine Ichor (Environment)
     [240448] = true, -- Quaking (Environment)
-    [342494] = true, -- Belligerent Boast (Season 1 Pridefull)
+    [342494] = true, -- Belligerent Boast (Season 1 Prideful)
     [-174773] = true, -- Spiteful Shade Melee Damage
     [350163] = true, -- Spiteful Shade Melee Damage
     [240446] = true, -- Explosion
+    [356414] = true, -- Frost Lance (Season 2)
+    [355806] = true, -- Massive Smash (Season 2)
+    [355737] = true, -- Scorching Blast (Season 2)
+    [358967] = true, -- Inferno (Season 2)
     -- Plaguefall
     [319898] = true, -- Vile Spit (Slime Tentacle)
     [328501] = true, -- Plague Bomb (Rigged Plagueborer)