+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Damage/Tabs/DamageDoneByPartyMembers/Tabs.lua
@@ -17,8 +17,9 @@ Creates tabs (with click-handlers) for Own Casts frame.
 function MyDungeonsBook:DamageDoneByPartyMembersFrame_CreateTabButtonsFrame(parentFrame, challengeId)
 	local tabs = self:TabsWidget_Create(parentFrame);
 	local tabsConfig = {};
+	local challenge = self:Challenge_GetById(challengeId);
 	for _, unitId in pairs(self:GetPartyRoster()) do
-		tinsert(tabsConfig, {value = unitId, text = self:GetNameByPartyUnit(challengeId, unitId)});
+		tinsert(tabsConfig, {value = unitId, text = self:GetUnitNameRoleStr(challenge.players[unitId])});
 	end
 	tabs:SetTabs(tabsConfig);
 	tabs:SetCallback("OnGroupSelected", function (container, _, tabId)