+++ UI/ChallengeDetails/Tabs/Roster/Tabs.lua
@@ -22,6 +22,7 @@ function MyDungeonsBook:RosterFrame_CreateTabButtonsFrame(parentFrame)
         {value = "equipment", text = L["Equipment"]},
         {value = "talents", text = L["Talents"]},
         {value = "covenant", text = L["Covenant"]},
+        {value = "misc", text = L["Misc"]},
     });
     tabs:SetCallback("OnGroupSelected", function (container, _, tabId)
         container:ReleaseChildren();
@@ -36,6 +37,9 @@ function MyDungeonsBook:RosterFrame_CreateTabButtonsFrame(parentFrame)
         end
         if (tabId == "talents") then
             self.talentsFrame = self:TalentsFrame_Create(container, self.activeChallengeId);
+        end
+        if (tabId == "misc") then
+            self.talentsFrame = self:MiscFrame_Create(container, self.activeChallengeId);
         end
     end);
     tabs:SetHeight(622);